Realized that difference between MvcControllerWithContext and WebApiControllerWithContext is only the template name and the logic for whether views need to be generated or not.
Since both the above are driven by the command line model easily, I don't see
a need for separate classes for the two, so merged into one class.

Now onto writing the template for WebAPI controller for EF context...
